desktop/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-multipart>=0.0.9
pywebview>=5.3.0
keyring>=25.0.0
python-docx>=1.1.0
aiofiles>=24.1.0
psutil>=5.9.0
